Save some garbage in the providers
