+++ src/main/java/org/mybatis/dynamic/sql/where/render/WhereClauseProvider.java
@@ -15,48 +15,10 @@
  */
 package org.mybatis.dynamic.sql.where.render;
 
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.Map;
-import java.util.Objects;
 
-public class WhereClauseProvider {
-    private final String whereClause;
-    private final Map<String, Object> parameters;
+public interface WhereClauseProvider {
+    Map<String, Object> getParameters();
 
-    private WhereClauseProvider(Builder builder) {
-        whereClause = Objects.requireNonNull(builder.whereClause);
-        parameters = Objects.requireNonNull(builder.parameters);
-    }
-
-    public Map<String, Object> getParameters() {
-        return Collections.unmodifiableMap(parameters);
-    }
-
-    public String getWhereClause() {
-        return whereClause;
-    }
-
-    public static Builder withWhereClause(String whereClause) {
-        return new Builder().withWhereClause(whereClause);
-    }
-
-    public static class Builder {
-        private String whereClause;
-        private final Map<String, Object> parameters = new HashMap<>();
-
-        public Builder withWhereClause(String whereClause) {
-            this.whereClause = whereClause;
-            return this;
-        }
-
-        public Builder withParameters(Map<String, Object> parameters) {
-            this.parameters.putAll(parameters);
-            return this;
-        }
-
-        public WhereClauseProvider build() {
-            return new WhereClauseProvider(this);
-        }
-    }
+    String getWhereClause();
 }
